fix delete

+++ admin/product/deleteProduct.php
@@ -1,6 +1,16 @@
 <?php
 	session_start();
-	include_once('connection.php');
+	$dbHost = 'localhost';
+	$dbUsername = 'TFT';
+	$dbPassword = 'Fongngu123';
+	$dbName = 'web_csdl';
+	$conn = mysqli_connect($dbHost,
+						$dbUsername,
+						$dbPassword,
+						$dbName);
+	if ($conn->connect_error) {
+		die("Connection failed: " . $conn->connect_error);
+	}
 
 	$productId = $_GET['productId'];
 	$sql = "DELETE FROM product WHERE productId = $productId";
@@ -9,5 +19,5 @@
 	} catch (mysqli_sql_exception) {
 		echo "Could not remove product!";
 	}
-	header('location:mainAdmin.php?page_layout=product');
+	header('location:..\mainAdmin.php?page_layout=product');
 ?>
